requirements.txt
# -----------------------------------------------
# Core Framework (GPU build - CUDA 11.8)
# -----------------------------------------------
#--index-url https://download.pytorch.org/whl/cu118
torch==2.7.1
huggingface_hub>=0.20.0 # for pushing to hub
# -----------------------------------------------
# Hugging Face Ecosystem
# -----------------------------------------------
transformers>=4.56.1
datasets>=4.0.0
evaluate[rouge]>=0.4.0
accelerate>=1.11.0
peft>=0.17.1
hf_transfer
huggingface_hub>=0.20.0 # for pushing to hub

# -----------------------------------------------
# Quantization and Memory-Efficient Training
# -----------------------------------------------
bitsandbytes>=0.42.0

# -----------------------------------------------
# Experiment Tracking and Configuration
# -----------------------------------------------
wandb>=0.22.0
python-dotenv>=1.1.1
PyYAML>=6.0.2

# -----------------------------------------------
# Optional but useful for reproducibility / plotting
# -----------------------------------------------
tqdm>=4.65.0
numpy>=1.26.0
matplotlib

# -----------------------------------------------
# OpenAI API
# -----------------------------------------------
openai>=2.7.0

# -----------------------------------------------
# Metrics
# -----------------------------------------------
rouge_score
